Delete test migrations in a workaround way.

+++ tests/conftest.py
@@ -100,6 +100,9 @@ def delete_migration_artifacts():
 
     for file in os.listdir(os.fsencode(rootdir)):
         filename = os.fsdecode(file)
+        if "e56a6f702357_create_table_checksum_and_logs" in filename:
+            continue
+
         if filename.endswith(".py"):
             os.remove(os.path.join(rootdir, filename))
         else:
@@ -121,7 +124,7 @@ class Client():
             raise RuntimeError("Need at least one schema dict for test client")
 
     def run_and_return_test_client(self):
-        # delete_migration_artifacts()
+        delete_migration_artifacts()
 
         self.initialize_test_client()
 
@@ -173,6 +176,7 @@ class Client():
             raise UpgradeFail
 
     def stop_container(self):
+        delete_migration_artifacts()
         self.postgres.stop_container()
